+++ HISTORY.md
@@ -6,6 +6,7 @@
 
 1. 修复Javascript插件Patch完数据后转回去可能类型错误的问题。
 2. 引入 `org.openjdk.nashorn` 和 `org.mozilla.rhino` 包，解决JDK 17以上没有javascript引擎的问题。
+3. 增加 `--default-field-separator` 选项允许设置Plain模式解析分隔符的默认值。
 
 ###